ok added the basic functionality to make fairshare work.  it doesn't do much yet, but it is ready to be called

+++ branches/priority/src/common/slurm_accounting_storage.h
@@ -215,7 +215,7 @@ typedef struct acct_association_rec {
 	uint32_t uid;		/* user ID */
 	
 	uint32_t used_jobs;	/* count of active jobs (DON'T PACK) */
-	uint32_t used_shares;	/* measure of resource usage */
+	uint64_t used_shares;	/* measure of resource usage */
 	uint32_t used_submit_jobs; /* count of jobs pending or running
 				    * (DON'T PACK) */
